this is the DbForms TAG LIB DESCRIPTOR file. it should be placed unter /WEB-INF directory of the Web Application.
NOTE for developers: this file is generated by transforming "taglib_docu.xml", using the XSL stylesheet "tld.xsl". (the latter was taken from JAKARTA-STRUTS project)
So, if you want to add a change to taglib.tld, you should apply it at first to taglib_docu.xml and then perform a xsl transformation on it. this way the docu and code are synchron. (with Servet api 2.3 this problem will be solved)