Patch by Bradley M. Handy (thanks), he wrote:
This patch addresses an assumption that the java executable has
an 'exe' extension which is not the case on *nix machines.

It also fixes a problem during the XSL translation where the
dependend/*.jar files are not used as the classpath.  Which results in a
ClassNotFoundException for those of us who don't have xalan.jar and
xerces.jar in our normal classpath.