Added the "defaultValue" attribute and the related accessor methods.
Now it's possible to specify a  "default value" for an image name.
For example, the name of your own "no-thumbnail-avaiable" icon ...

Example:

<db:blobURL fieldName="THUMBNAIL" defaultValue="myImage.jpg"/>

can be used with ->

<img src="<db:blobURL fieldName="THUMBNAIL" defaultValue="myImage.jpg"/>" border="0" width="xx"
height="yy">

In that manner, when the thumbnail image name is not
avaiable (the related db field is blank), the FileServlet
uses the specified "defaultValue attribute" value as the
image name.

See also the related changes into Fileservlet class