Contributed by Shawn:
This version is a merge of the current taglib_docu.xml
and the information within reference part of old dbforms
user guide. It does not add, delete or alter any tag
or attribute, but inserts lots of comments that were
previously only within user guide.

So the generated tld will be the same, but the generated
documentation will (hopefully) be better. The document
might need a bit reformating now, but should already be
usable.