New attribute 'showAlwaysInFooter' (default: "true")  for tag NavNewButton to get following behaviour:
A NavNewButton is different from other buttons, it will normally
even been shown if it is placed within footer of a dbforms tag
with empty result set.  This makes sense if you want to show a
set of rows and then below always the button to open
a form to insert a new row. But if you have a form with body
attribute allowNew set to "true", an insert form will automatically
be shown in case no row is found. In that case the button might
irritate users, you should then set attribute showAlwaysInFooter
to "false".  Then the button will not be shown if result set is empty.