fix the
String getEventIdFromDestinationTable(HttpServletRequest request, String action)

method to accept events generated by db:gotoButton tag.

Here's the explanation of the bug from Henner Kollmann:

"Know i tried to use the db:gotoButton. In this case action is something
like ac_goto_11 which is the number of the button! Know in current
implemtation you will try to get the tableId from directly from the
action - because it is not a GET event. This will result in table 11,
which is definitively wrong because it is the number of the button."