refactoring with EXTRACT METHOD:
created the new private processInvolvedTables()

extracting code from the main process() method.

I used the the testsuite and bookstore test suite with the previous CVS
version and with this one.
It seems the behaviour of the controller is ok.

PLEASE, could you make some tests and let me know if this version is ok for you.